backend/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.20
psycopg2-binary==2.9.6
requests==2.31.0
python-dotenv==1.0.0
geographiclib==2.0.4
